Remove Deepak Prabhakara from list of Maintainers
As @deepakprabhakara is no longer with Red Sift, and hasn't contributed to the project throughout his membership, I propose the removal of his maintainer rights.

Of course, we welcome any further contributions, and we may consider extending the maintainer rights again as we see fit.

This proposition requires 1 approval to be merged by either @deepakprabhakara or @rhdxmr .

Signed-off-by: rsdy <p@symmetree.dev>